bcmdhd: only append firmware revision during firmware loading
Inorder to find the revision of the chip and concatenate the firmware
filename, the BUS should be active so that the chip revision detection
could take place. This is failing during resets when the bus is inactive.
Move the detection logic to happen during the firmware downloading at
which point we can be sure that the bus is active.

Change-Id: I15cfd6197d537ff96e45b5ac5375ff76ad678ea4