i9300: defconfig: disable samsung_extdisp driver
The samsung_extdisp driver for GT-I9500 was affected by CVE-2015-1800 and
CVE-2015-1801. The Galaxy S3 uses an older version of this driver and is at
least affected by CVE-2015-1801. Newer kernel versions for the GT-I9500
had the driver completely removed. I also found no indication that the
driver is actually needed for the S3, so let's disable it.

vulnerability disclosure and further information:
http://blog.quarkslab.com/kernel-vulnerabilities-in-the-samsung-s4.html

Change-Id: Ia4b484026bde1145173f8765e1e3c8957576585d
Signed-off-by: Wolfgang Wiedmeyer <wolfgit@wiedmeyer.de>